pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "walle-bot"
version = "0.1.0"
description = "Telegram moderation bot"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==22.6",
    "PyYAML==6.0.2",
    "python-dotenv==1.1.1",
]

[project.optional-dependencies]
dev = ["pytest==8.4.2"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
walle-bot = "walle_bot.main:main"